requirements.txt
pygame-ce
pygbag
auto-py-to-exe
